+++ arch/arm/src/lc823450/lc823450_dvfs2.c
@@ -131,7 +131,7 @@ static uint64_t _get_current_time64(void)
   struct timespec ts;
 
   clock_systime_timespec(&ts);
-  return (uint64_t)ts.tv_sec * NSEC_PER_SEC + (uint64_t)ts.tv_nsec;
+  return ts.tv_sec * NSEC_PER_SEC + ts.tv_nsec;
 }
 
 /****************************************************************************
