+++ arch/arm/src/stm32f7/stm32_tickless.c
@@ -973,8 +973,8 @@ int up_timer_start(const struct timespec *ts)
 
   /* Express the delay in microseconds */
 
-  usec = (uint64_t)ts->tv_sec * USEC_PER_SEC +
-         (uint64_t)(ts->tv_nsec / NSEC_PER_USEC);
+  usec = ts->tv_sec * USEC_PER_SEC +
+         (ts->tv_nsec / NSEC_PER_USEC);
 
   /* Get the timer counter frequency and determine the number of counts need
    * to achieve the requested delay.
@@ -1021,7 +1021,7 @@ int up_timer_start(const struct timespec *ts)
 int up_alarm_start(const struct timespec *ts)
 {
   size_t offset = 1;
-  uint64_t tm = ((uint64_t)ts->tv_sec * NSEC_PER_SEC + ts->tv_nsec) /
+  uint64_t tm = (ts->tv_sec * NSEC_PER_SEC + ts->tv_nsec) /
                 NSEC_PER_TICK;
   irqstate_t flags;
