+++ arch/xtensa/src/esp32/esp32_oneshot.c
@@ -274,8 +274,8 @@ int esp32_oneshot_start(struct esp32_oneshot_s *oneshot,
 
   /* Retrieve the duration from timespec in microsecond */
 
-  timeout_us = (uint64_t)ts->tv_sec * USEC_PER_SEC +
-               (uint64_t)(ts->tv_nsec / NSEC_PER_USEC);
+  timeout_us = ts->tv_sec * USEC_PER_SEC +
+               (ts->tv_nsec / NSEC_PER_USEC);
 
   /* Verify if it is a multiple of the configured resolution.
    * In case it isn't, warn the user.
@@ -443,7 +443,7 @@ int esp32_oneshot_current(struct esp32_oneshot_s *oneshot, uint64_t *usec)
 
   ESP32_TIM_GETCTR(oneshot->tim, usec);
 
-  *usec = *usec * (uint64_t)oneshot->resolution;
+  *usec = *usec * oneshot->resolution;
 
   return OK;
 }
