+++ arch/xtensa/src/esp32s3/esp32s3_tim_lowerhalf.c
@@ -410,7 +410,7 @@ static int timer_lh_settimeout(struct timer_lowerhalf_s *lower,
 
   /* Set the timeout */
 
-  ESP32S3_TIM_SETALRVL(priv->tim, (uint64_t)timeout);
+  ESP32S3_TIM_SETALRVL(priv->tim, timeout);
 
   return ret;
 }
