+++ drivers/note/notesnap_driver.c
@@ -398,13 +398,13 @@ void notesnap_dump_with_stream(FAR struct lib_outstream_s *stream)
 
       perf_convert(note->count, &time);
       lib_sprintf(stream,
-                  "snapshoot: [%" PRIu64 ".%09u] "
+                  "snapshoot: [%" PRIu64 ".%09ld] "
 #ifdef CONFIG_SMP
                   "[CPU%d] "
 #endif
                   "[%d] %-16s %#" PRIxPTR "\n",
-                  (uint64_t)time.tv_sec,
-                  (unsigned)time.tv_nsec,
+                  time.tv_sec,
+                  time.tv_nsec,
 #ifdef CONFIG_SMP
                   note->cpu,
 #endif
